+++ packages/article-byline/article-byline.stories.js
@@ -1,13 +1,15 @@
 import React from "react";
 import { Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
+import { checkA11y } from "@storybook/addon-a11y";
+import { fonts } from "@times-components/styleguide";
 import ArticleByline from "./article-byline";
 
 const authorsAST = require("./fixtures/authors.json");
 
 const bylineStyles = {
   fontSize: 13,
-  fontFamily: "GillSansMTStd-Medium",
+  fontFamily: fonts.supporting,
   color: "#696969",
   lineHeight: 13,
   flexDirection: "row"
@@ -21,6 +23,7 @@ const bylineLinkStyles = {
 };
 
 storiesOf("Primitives/ArticleByline", module)
+  .addDecorator(checkA11y)
   .add("ArticleByline with a single author", () => (
     <Text style={bylineStyles}>
       <ArticleByline ast={authorsAST.singleAuthor} />