+++ packages/depend/pat-regexp.js
@@ -1,0 +1,7 @@
+export default function patRegExp(pat) {
+  if (pat === "") {
+    return "";
+  }
+
+  return `^${pat.replace(/[*]/g, ".+")}.*$`;
+}