+++ packages/icons/icons.stories.js
@@ -1,7 +1,10 @@
 import React from "react";
 import { View, StyleSheet, Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
-import { IconDiamond, IconTwitter } from "./icons";
+import { checkA11y } from "@storybook/addon-a11y";
+import { fonts } from "@times-components/styleguide";
+import { color } from "@storybook/addon-knobs";
+import { IconDiamond, IconTwitter, IconVideo } from "./icons";
 
 const borderWidth = 1;
 const borderColor = "#d3d3d3";
@@ -29,19 +32,38 @@ const styles = StyleSheet.create({
   },
   label: {
     paddingTop: 15,
-    fontFamily: "GillSansMTStd-Medium"
+    fontFamily: fonts.supporting
   }
 });
 
-storiesOf("Primitives/Icons", module).add("Icons", () => (
+
+storiesOf("Primitives/Icons", module)
+  .addDecorator(checkA11y)
+  .add("Icons", () => (
   <View style={styles.wrapper}>
     <View style={styles.icon}>
-      <IconDiamond width={50} height={50} fillColour="#4D4D4D" />
+      <IconDiamond
+        width={50}
+        height={50}
+        fillColour={color("Diamond color", IconDiamond.defaultProps.fillColour)}
+      />
       <Text style={styles.label}>Diamond</Text>
     </View>
     <View style={styles.icon}>
-      <IconTwitter width={50} height={50} />
+      <IconTwitter
+        width={50}
+        height={50}
+        fillColour={color("Icon color", IconTwitter.defaultProps.fillColour)}
+      />
       <Text style={styles.label}>Twitter</Text>
+    </View>
+    <View style={styles.icon}>
+      <IconVideo
+        width={50}
+        height={50}
+        fillColour={color("Video color", IconVideo.defaultProps.fillColour)}
+      />
+      <Text style={styles.label}>Video</Text>
     </View>
   </View>
 ));