+++ packages/icons/icons/proptypes.js
@@ -1,0 +1,7 @@
+import PropTypes from "prop-types";
+
+export default {
+  fillColour: PropTypes.string,
+  height: PropTypes.number.isRequired,
+  width: PropTypes.number.isRequired
+};