+++ packages/image/image.stories.web.js
@@ -3,6 +3,7 @@ import React from "react";
 import ReactDOMServer from "react-dom/server";
 import { View } from "react-native";
 import { storiesOf } from "@storybook/react-native";
+import { checkA11y } from "@storybook/addon-a11y";
 import Image from "./image";
 
 const squareUri =
@@ -11,9 +12,9 @@ const SquareImage = props => (
   <Image uri={squareUri} aspectRatio={1 / 1} {...props} />
 );
 
-storiesOf("Primitives/Image", module).add(
-  "Server side rendered Image (web only)",
-  () => {
+storiesOf("Primitives/Image", module)
+  .addDecorator(checkA11y)
+  .add("Server side rendered Image (web only)", () => {
     const markup = {
       __html: ReactDOMServer.renderToStaticMarkup(
         <View>
@@ -46,5 +47,4 @@ storiesOf("Primitives/Image", module).add(
     };
 
     return <div dangerouslySetInnerHTML={markup} />;
-  }
-);
+  });