+++ packages/slice/templates/default/responsive.web.js
@@ -16,8 +16,8 @@ export const ChildContainer = withResponsiveStyles(View, {
 });
 ChildContainer.displayName = "ChildContainer";
 
-export const getChildrenContainer = childCount =>
-  withResponsiveStyles(View, {
+export const getChildrenContainer = ({ childCount }) => {
+  const ChildrenContainer = withResponsiveStyles(View, {
     base: () => `
     flex: 1;
     flex-direction: column;
@@ -32,6 +32,8 @@ export const getChildrenContainer = childCount =>
     `;
 
       const largeStyle = `
+      padding-left: 20px;
+      padding-right: 20px;
       width: 100%;
     `;
 
@@ -44,3 +46,6 @@ export const getChildrenContainer = childCount =>
     width: ${childCount >= 3 ? "100%" : config.wideBpWidth};
   `
   });
+  ChildrenContainer.displayName = "ChildrenContainer";
+  return ChildrenContainer;
+};