+++ packages/styleguide/styleguide.js
@@ -3,6 +3,8 @@ import functionalColours from "./colours/functional";
 
 import FadeIn from "./animations";
 
+import timesFonts from "./fonts";
+
 export const colours = {
   section: sectionColours,
   functional: functionalColours
@@ -11,3 +13,5 @@ export const colours = {
 export const Animations = {
   FadeIn
 };
+
+export const fonts = timesFonts;