+++ packages/tracking/__tests__/tracking-context.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Text } from "react-native";
+import { Text, View } from "react-native";
 import PropTypes from "prop-types";
 import renderer from "react-test-renderer";
 import trackingContextTypes from "../tracking-context-types";
@@ -274,5 +274,40 @@ module.exports = () => {
 
       expect(WithTrackingContext.someStatic).toEqual(TestComponent.someStatic);
     });
+
+    it("tracks nested events", () => {
+
+      const Test1 = props => <div> {props.children} </div>;
+      const Test2 = props => <span> {props.b} </span>;
+      
+      Test1.propTypes = { a: PropTypes.string, children: PropTypes.node };
+      Test2.propTypes = { b: PropTypes.string };
+
+
+      const Tracker1 = withTrackingContext(
+        Test1, {
+          trackingObjectName: "tracker1",
+          getAttrs: ({a}) => ({a})
+        }
+      );
+
+      const Tracker2 = withTrackingContext(
+        Test2, {
+          trackingObjectName: "tracker2",
+          getAttrs: ({b}) => ({b})
+        }
+      );
+
+      const reporter = jest.fn();
+
+      renderer.create(
+        <Tracker1 a="a" analyticsStream={reporter}>
+          <Tracker2 b="b"/>
+        </Tracker1>
+      );
+
+      expect(reporter.mock.calls).toEqual({});
+    });
+
   });
 };