+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.12"></a>
+## [0.12.12](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.11...@times-components/article-byline@0.12.12) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
+<a name="0.12.11"></a>
+## [0.12.11](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.10...@times-components/article-byline@0.12.11) (2018-02-28)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-byline
+
 <a name="0.12.10"></a>
 ## [0.12.10](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.12.9...@times-components/article-byline@0.12.10) (2018-02-26)