+++ packages/date-publication/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.12"></a>
+## [0.11.12](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.11.11...@times-components/date-publication@0.11.12) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/date-publication
+
+<a name="0.11.11"></a>
+## [0.11.11](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.11.10...@times-components/date-publication@0.11.11) (2018-02-28)
+
+
+
+
+**Note:** Version bump only for package @times-components/date-publication
+
 <a name="0.11.10"></a>
 ## [0.11.10](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.11.9...@times-components/date-publication@0.11.10) (2018-02-23)