+++ packages/image/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.13.25"></a>
+## [1.13.25](https://github.com/newsuk/times-components/compare/@times-components/image@1.13.24...@times-components/image@1.13.25) (2018-03-01)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
+<a name="1.13.24"></a>
+## [1.13.24](https://github.com/newsuk/times-components/compare/@times-components/image@1.13.23...@times-components/image@1.13.24) (2018-02-28)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
+<a name="1.13.23"></a>
+## [1.13.23](https://github.com/newsuk/times-components/compare/@times-components/image@1.13.22...@times-components/image@1.13.23) (2018-02-28)
+
+
+
+
+**Note:** Version bump only for package @times-components/image
+
 <a name="1.13.22"></a>
 ## [1.13.22](https://github.com/newsuk/times-components/compare/@times-components/image@1.13.21...@times-components/image@1.13.22) (2018-02-26)